refactor(): move test files to test directory
